scripts: extract_dts_includes.py: Support linux pinctrl
Extends the extraction script to generate define directives for pinctrl
properties specified in https://www.kernel.org/doc/Documentation/devicetree\
/bindings/pinctrl/pinctrl-bindings.txt to support the migration of pinmux
drivers and other drivers to device tree usage.

Does not change the generation for pinctrl bindings not specified in linux
pinctrl-bindings.txt. As the Linux pinctrl property names are not in use in
the current codebase there is no impact on it.

The generation of define directives is extended in the following ways:

Array/ list property values
---------------------------

Array/ list property values are generated as a comma separated list v1,v2,..
(used to be [v1,v2,...]). This helps to use the define in macros and functions.

Pin configurations specified by pinctrl-x
-----------------------------------------

The generated identifier is reduced to the bare minimum to make it unique.
Device tree changes that do not affect the handling of the configuration do not
change the identifier by unneeded information.

e.g.

pincontroller@40000000 {
    device_rx_port_a@000 {
        rx {
            pinmux = <PORT_A0, PINMUX_RX_A0>;
        };
    };
    device_rx_port_b@001 {
        rx {
            pinmux = <PORT_B1, PINMUX_RX_B1>;
        };
    };
    label = "pinmux";
};

&device {
    pinctrl-0 = <&device_rx_port_a>;
    pinctrl-names = "default";
    status = "ok";
};

creates:

#define DEVICE_xx_PINCTRL_RX_DEFAULT_PINMUX     PORT_A0,PINMUX_RX_A0
                                     ------  -> configuraton property name
                             -------         -> pinctrl name (of pinctrl-x)
                          --                 -> configuration identification
        ---------                            -> default prefix (no change)

if port B is selected (instead of port A):

&device {
    pinctrl-0 = <&device_rx_port_b>;
    pinctrl-names = "default";
    status = "ok";
};

the same identifier is created (but a different token-string):

#define DEVICE_xx_PINCTRL_DEFAULT_RX_PINMUX     PORT_B1,PINMUX_RX_B1

Default pin configurations specified by pinctrl-x and pinctrl-names
-------------------------------------------------------------------

For default pin configurations (pinctrl name is "default") additional define
directives are generated. These define directives can be used by pinmux drivers
to initialise the pin at startup.

e.g. if port A is selected:

#define PINMUX_4000000_PINCTRL_000_PINMUX     PORT_A0,PINMUX_RX_A0
                                   ------  -> configuraton property name
                               ---         -> configuration index
        --------------                     -> default prefix of pin controller

e.g. if port B is selected

#define PINMUX_4000000_PINCTRL_000_PINMUX     PORT_B1,PINMUX_RX_B1

The configuration index is created on the fly. It makes the identifiers unique
and independent from the specific naming of the node.
Such pinmux drivers can assume a sequential set of default initialisations.
No additional mapping of the identifiers by the driver is necessary.

The default configuration define directives for all linux pinctrl-bindings are
generated. All configuration values are initially set to 0. The value is
replaced by "real" values for all properties given in the device tree. This way
pinmux drivers can rely on all macros defined.

The total number of default configurations generated is given by:

#define PINMUX_4000000_PINCTRL_COUNT   xxx

The unit test shows samples for pinmux default initialisation using the
macros.

The change was verified by sanitycheck. There were several warnings about
"WARNING: ram_size", "WARNING: rom_size" which seem not related to the change.
No faults were detected.

A unit test for the script was added. Tests are passed !-).

Signed-off-by: Bobby Noelte <b0661n0e17e@gmail.com>