scripts: extract_dts_includes.py: new pinctrl and related directives
The extract_dts_includes.py is extended to extract new directives.
The pinctrl and gpio-ranges directives as described in the Linux
pinctrl-bindings.txt. The <device>-controller directive as commonly
used by several nodes (e.g. gpio-controller, pin-controller, ...)
and the compatible directive.

The extraction and define generation for current pinctrl directives
is unchanged.

Pin controller nodes
--------------------

For a pin controller node five data sets are extracted from the device
tree and provided in generated_dts_board.h.

- FUNCTION

For all active client nodes that reference a pin-controller by
pinctrl-x(&pinctrl, ...) FUNCTION defines are generated.
The function equals the client node.

- STATE_NAME

For all pinctrl names of pinctrl-x directives of active client nodes
STATE_NAME defines are generated. This is mostly to allow deduplication
of commonly used state names such as e.g. “default”.

- PINCTRL_STATE

For all pinctrl-x directives of active client nodes PINCTRL_STATE
defines are generated. The pinctrl state references the FUNCTION
(aka. node) it is given in. The pinctrl state also references the
STATE_NAME given in the pinctrl-names directive associated to the
pinctrl-x directive.

All the pins that a pinctrl state controls are regarded to be a pin
group. The name of the pin group is the same as the name of the
pinctrl state.

- PINCTRL

For all subnodes of pin configurations referenced by active client nodes
- pinctrl-x(&xxx, pinconf_a, ...) - PINCTRL defines are generated. The
pinctrl references the FUNCTION (aka. node) it is referenced from. The
pinctrl also references the PINCTRL_STATE it is referenced from.

- GPIO_RANGE

For all gpio-ranges directives of active GPIO nodes GPIO_RANGE defines
are generated.

Pin controller client nodes
---------------------------

For client nodes of pin controller nodes two datasets are extracted from
the device tree and provided in generated_dts_board.h.

- PINCTRL_CONTROLLER

For all pin controllers that are referenced in a pinctrl-x directive
PINCTRL_CONTROLLER defines are generated.

- PINCTRL

For all subnodes of pin configurations referenced by active client nodes
- pinctrl-x(&xxx, pinconf_a, ...) - PINCTRL defines are generated.
The pinctrl references the PINCTRL_CONTROLLER given in the pinctrl-x
directive. The pinctrl also references the PINCTRL_STATE the pin
controller associates to this pinctrl.

GPIO nodes
----------

- GPIO_RANGE

For all gpio-ranges directives of active GPIO nodes GPIO_RANGE defines
are generated.

Signed-off-by: Bobby Noelte <b0661n0e17e@gmail.com>