drivers: pinctrl: add pinctrl driver library
Add a header-only library for pinctrl drivers.

The library abstracts away the complex generation of the
generic parts of a pinctrl driver such as:
- compile time initialisation of configuration data from device tree
- access functions to config data
- compile time initialisation of whole driver structs

The driver implementer does only have to implement the necessary
hardware access routines.

The library supports up to 4 pin controller devices to be
instantiated.

The pinctrl_lib uses the Chaos preprocessor library to accomplish
more complex C macro processing. This is fully hidden from the
user. A driver implementer must only define a set of library
configuration defines.

Signed-off-by: Bobby Noelte <b0661n0e17e@gmail.com>