tests: drivers: gpio: support NUCLEO F091RC and F030R8 boards
The NUCLEO FO91RC and F030R8 boards use the GPIO driver with PINCTRL
backend by default. The STM32F0 SoC on the board does not support
level interrupts.

The test is adapted to exclude the level interrupt tests and
to correctly set up the GPIO pins for the board and the choosen
GPIO driver with PINCTRL backend.

Signed-off-by: Bobby Noelte <b0661n0e17e@gmail.com>