codegen: Rewrite config.py
There were several issues with config.py that have now been rectified
in this rewrite.

Scripts should not be hardcoding paths to generated config files. It
is the responsibility of the build system to invoke scripts and to
feed them with information about where files are located.

Scripts should avoid using environment variables. Instead information
should be explicitly passed from the build system to the
scripts. Being explicit instead of implicit makes it easier to keep
track of dependencies between modules.

config.py has been parsing a header file (autoconf.h), but Kconfig
fragments have a simpler syntax and are therefore easier to parse (No
need for shlex).

config.py used the string Type for all extracted values, but it is
more convenient for users to have access to the Python-equivalent
type.

All of these issues have been resolved in the rewrite.

Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>