templates: add pinctrl driver template
templates/drivers/pinctrl_tmpl.c:
The template abstracts away the complex generation of the
generic parts of a pinctrl driver such as:
- compile time initialisation of configuration data from device tree
- access functions to config data
- compile time initialisation of whole driver structs

The driver implementer does only have to implement the necessary
hardware access routines.

The template is to be used by inline code generation.
A driver implementer must only define a set of python
global variables to configure the code generation.

Signed-off-by: Bobby Noelte <b0661n0e17e@gmail.com>