build: fix support of out of tree serial driver
The Zephyr serial driver lib depends on CONFIG_SERIAL_HAS_DRIVER
but this is an indication that there is a serial driver. An out
of tree driver providing this indication triggers the Zephyr
serial driver lib to be build with no driver which leads to
an error.

Make the Zephyr driver lib depend on CONFIG_SERIAL which is also
the semantically correct dependency.

Signed-off-by: Bobby Noelte <b0661n0e17e@gmail.com>